Remove backup page
