+++ packages/browser-wallet/src/popup/pages/VerifiableCredential/i18n/en.ts
@@ -6,7 +6,8 @@ const t = {
     menu: {
         revoke: 'Revoke',
         details: 'Details',
-        backup: 'Backup',
+        import: 'Open import window',
+        export: 'Download export file',
     },
     details: {
         id: 'Credential holder ID',
