+++ packages/browser-wallet/src/popup/pages/VerifiableCredentialBackup/VerifiableCredentialBackup.scss
@@ -1,13 +1,5 @@
 $header-height: rem(56px);
 
-.verifiable-credential-backup {
-    display: flex;
-    justify-content: space-around;
-    flex-direction: column;
-    height: calc(100% - 56px);
-    background-color: $color-bg;
-}
-
 .verifiable-credential-import {
     display: flex;
     flex-direction: column;
