+++ packages/browser-wallet/src/popup/pages/VerifiableCredentialBackup/i18n/en.ts
@@ -1,15 +1,6 @@
 export default {
-    import: {
-        title: 'Import Web3 ID Credentials',
-        noImported: 'No Web3 ID Credentials were imported',
-        error: 'Unable to import the chosen file. The file must be a backup created with the same seed phrase.',
-    },
-    backup: {
-        header: 'Web3 ID Credentials backup',
-        button: {
-            import: 'Go to import page',
-            export: 'Download export file',
-        },
-    },
+    title: 'Import Web3 ID Credentials',
+    noImported: 'No Web3 ID Credentials were imported',
+    error: 'Unable to import the chosen file. The file must be a backup created with the same seed phrase.',
     close: 'Close',
 };
